+++ homematicip/aio/device.py
@@ -462,11 +462,19 @@ class AsyncAccelerationSensor(AccelerationSensor, AsyncDevice):
         )
 
 
-class AsyncGarageDoorModuleTormatic(GarageDoorModuleTormatic, AsyncDevice):
-    """ HMIP-MOD-TM (Garage Door Module Tormatic) """
+class AsyncDoorModule(DoorModule, AsyncDevice):
+    """ Generic Door Module class """
 
     async def send_door_command(self, doorCommand=DoorCommand.STOP):
         return await self._connection.api_call(*super().send_door_command(doorCommand))
+
+
+class AsyncGarageDoorModuleTormatic(GarageDoorModuleTormatic, AsyncDoorModule):
+    """ HMIP-MOD-TM (Garage Door Module Tormatic) """
+
+
+class AsyncHoermannDrivesModule(HoermannDrivesModule, AsyncDoorModule):
+    """ HMIP-MOD-HO (Garage Door Module for Hörmann) """
 
 
 class AsyncPluggableMainsFailureSurveillance(