+++ homematicip/class_maps.py
@@ -29,6 +29,7 @@ TYPE_CLASS_MAP = {
     DeviceType.HEATING_SWITCH_2: HeatingSwitch2,
     DeviceType.HEATING_THERMOSTAT: HeatingThermostat,
     DeviceType.HEATING_THERMOSTAT_COMPACT: HeatingThermostatCompact,
+    DeviceType.HOERMANN_DRIVES_MODULE: HoermannDrivesModule,
     DeviceType.KEY_REMOTE_CONTROL_4: KeyRemoteControl4,
     DeviceType.KEY_REMOTE_CONTROL_ALARM: KeyRemoteControlAlarm,
     DeviceType.LIGHT_SENSOR: LightSensor,